# public/swagger.yaml
openapi: 3.0.3
info:
  title: Food Delivery System API
  description: |
    RESTful API for Food Delivery System with RAG Chatbot

    ## Features
    - User authentication & authorization
    - Restaurant & menu management
    - Order processing & tracking
    - Real-time delivery tracking
    - Push notifications
    - Chat messaging
    - RAG-powered chatbot

    ## Base URL
    - Development: `http://localhost:3000`
  version: 1.0.0
  contact:
    name: API Support
    email: support@fooddelivery.com
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: customer@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      name:
                        type: string
                      role:
                        type: string
                        enum: [CUSTOMER, RESTAURANT_OWNER, DRIVER, ADMIN]
...
  /api/menu:
    get:
      tags:
        - Menu
      summary: Get menu items
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: string
        - name: menuItemId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of menu items
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
...
  /api/favorites:
    get:
      tags:
        - Favorites
      summary: Get user favorites
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of favorite restaurants

    post:
      tags:
        - Favorites
      summary: Add restaurant to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                restaurantId:
                  type: string
      responses:
        '201':
          description: Favorite added
  /api/chat/messages:
    get:
      tags:
        - Chat
      summary: Get chat messages
      parameters:
        - name: orderId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of messages

    post:
      tags:
        - Chat
      summary: Send chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                senderId:
                  type: string
                receiverId:
                  type: string
                message:
                  type: string
      responses:
        '201':
          description: Message sent
  /api/chatbot:
    post:
      tags:
        - Chatbot
      summary: Ask chatbot (RAG)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  example: แนะนำร้านอาหารหน่อย
      responses:
        '200':
          description: Chatbot response
components:
  schemas:
...
    MenuItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        isAvailable:
          type: boolean
...